+++ lib/monologue.rb
@@ -26,6 +26,7 @@ module Monologue
                  :admin_force_ssl,
                  :posts_per_page,
                  :google_analytics_id,
+                 :gauge_analytics_site_id,
                  :layout,
                  :sidebar
 end